Add a password_scheme field in mailbox table.
This is useful if you want  to deal with multiple password scheme. The
one you use today might be too  weak in the future and you should have
the possibility to change it.

The password scheme names stored in this new field are the one
described here :
http://wiki2.dovecot.org/Authentication/PasswordSchemes

To use it with Dovecot, you just have to do a query like :
SELECT CONCAT('{', password_scheme, '}', password) FROM `mailbox`
WHERE [...]

The password_scheme  field will contain  an empty string  for password
which have been set before this change.