Implementation of Dovecot's salted SHA password schemes
Updated Dovecot.php to provide a generate and check password function and
implemented support for the SSHA* algorithms of Dovecot. For those 3 algos,
no shell access is required, the others fail back.

Mailbox.php provides a function to check a password. For dovecote SHA-schemes,
the new check_password Function in Dovecot.php is used. Others are handled
as before.

AuthController relies on the new check function located in the Mailbox.php