[BF] Do not soft-delete mailboxes when not deleting the file system
A new feature allowing the on disk deletion of mailboxes created an issue when purging mailboxes without deleting the file system in that the mailbox record was marked as soft-deleted rather than actually deleted. This prevented the same mailbox from being added again in the future. This is corrected in this release but it now requires that on disk deletion be explicitly enabled in application.ini by adding:

mailbox_deletion_fs_enable = true