Added table of contents and reordered email settings text to make it more clear (#204)
* Text summary and reorder

Added a little table of contents to the email text to make  sure people
are aware of the full contents of the email and reordered the "check
settings below" to immediately before the settings. By having it before
the change password makes it less prone to be read and correctly
understood

* Revert "Text summary and reorder"

This reverts commit 3d6623b3f70276b4e2d3edae856212f24eefb2f8.

* Table of contents and reorder

Added a table of contents to the email text as it may not be clear the
email full length, and reorder the line to check settings below to
immediately before the settings, as it could get easily forgotten and
misunderstood if staying before the change password.